+++ test/core/k_report_generation_test.py
@@ -45,7 +45,7 @@ def test_create_report():
                  'html_links': [{'path': str(test_dir),
                                  'name': 'index.html',
                                  'label': 'index.html',
-                                 'description': 'HTML report for GTDBTk Classify'
+                                 'description': 'GTDB-Tk Classify Report'
                                  }],
                  'file_links': [{'shock_id': shock_id,
                                  'name': 'GTDB-Tk_classify_wf.zip',